Changed goods price type to int64

+++ migrations/20250413181139_create_goods_table.sql
@@ -3,7 +3,7 @@
 CREATE TABLE IF NOT EXISTS goods (
     number VARCHAR(255) NOT NULL,
     description VARCHAR(255) NOT NULL,
-    price DOUBLE PRECISION NOT NULL,
+    price INT64 NOT NULL,
     updated_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP
 );
 -- +goose StatementEnd
